Rename keyboard_handler to avoid naming conflict
